First functional version of MCASlitegit status
git-svn-id: http://192.168.1.2:8080/svn/mcaslite2.0@5 375a5ec8-f457-43dc-b3d4-3e37c4652669

+++ manager/trunk/src/cat/i2cat/mcaslite/utils/MediaUtils.java
@@ -2,9 +2,11 @@ package cat.i2cat.mcaslite.utils;
 
 import java.io.File;
 import java.net.URI;
+import java.util.List;
 
 import org.apache.commons.io.FileUtils;
 import cat.i2cat.mcaslite.entities.ApplicationConfig;
+import cat.i2cat.mcaslite.entities.Transco;
 import cat.i2cat.mcaslite.exceptions.MCASException;
 
 public class MediaUtils {
@@ -31,7 +33,11 @@ public class MediaUtils {
 	public static void toDestinationUri(String file, URI uri) throws MCASException {
 		try {
 			if (uri.getScheme().equals("file")) {
-				FileUtils.copyFile(new File(file), new File(uri.getPath()));
+				if ((new File(file)).exists()) {
+					FileUtils.copyFile(new File(file), new File(uri.getPath()));
+				} else {
+					throw new MCASException();
+				}
 			} else if (uri.getScheme().equals("http")) {
 				//TODO
 			} else if (uri.getScheme().equals("https")) {
@@ -46,11 +52,28 @@ public class MediaUtils {
 			throw new MCASException();
 		}
 	}
+	
+	public static void deleteFile(String file){
+		File fd = new File(file);
+		if (fd.exists()){
+			fd.delete();
+		}
+	}
 
-	public static void deleteInput(String id) {
-		File file = new File(ApplicationConfig.getInputWorkingDir() + id);
-		if (file.exists()){
-			file.delete();
+	private static void cleanTransco(Transco transco){
+		File input = new File(transco.getInputFile());
+		File output = new File(transco.getOutputFile());
+		if (input.exists()){
+			input.delete();
+		}
+		if (output.exists()){
+			output.delete();
+		}
+	}
+	
+	public static void clean(List<Transco> transcos){
+		for(Transco transco : transcos){
+			cleanTransco(transco);
 		}
 	}
